Added linting

+++ export_images/export_images.ts
@@ -1,68 +1,68 @@
 function displayMessage(text: string): void {
-  var messageWindow = new Window("dialog", undefined, undefined, {
-    borderless: true
+  const messageWindow = new Window('dialog', undefined, undefined, {
+    borderless: true,
   });
-  messageWindow.add("statictext", undefined, text);
-  messageWindow.add("button", undefined, "OK");
+  messageWindow.add('statictext', undefined, text);
+  messageWindow.add('button', undefined, 'OK');
   messageWindow.show();
 }
 
 function getActiveDocument(app: Application): Document {
-  if(app.activeDocument) {
-    return app.activeDocument
+  let result: Document;
+
+  if (app.documents.length < 0) {
+    displayMessage('There are no opened documents');
+    result = null;
   } else {
-    prompt("There are no opened documents")
+    result = app.activeDocument;
   }
+
+  return result;
 }
 
 function getDestinationPath(pathToDocument: string): string {
-  let newPath: string;
-  let name: string;
+  const name: string = pathToDocument.split('/').pop();
+  const newPath: string = `${pathToDocument.slice(0, -name.length)}_media`;
 
-  name = pathToDocument.split("/").pop();
-  newPath = `${pathToDocument.slice(0, - name.length)}_media` ;
+  const checkFolder: Folder = new Folder(`${newPath}/`);
 
-  let checkFolder: Folder = new Folder(`${newPath}/`);
-
-  if (checkFolder.exists) {
-    return newPath;
-  } else {
+  if (!checkFolder.exists) {
     checkFolder.create();
-    return newPath;
   }
+
+  return newPath;
 }
 
 
 function createImages(document: Document, destPath: string, filename: string): void {
-
-  let options: ExportForScreensOptionsJPEG = new ExportForScreensOptionsJPEG;
+  const options: ExportForScreensOptionsJPEG = new ExportForScreensOptionsJPEG();
   options.compressionMethod = JPEGCompressionMethodType.BASELINESTANDARD;
   options.embedICCProfile = false;
   options.antiAliasing = AntiAliasingMethod.TYPEOPTIMIZED;
   options.scaleType = ExportForScreensScaleType.SCALEBYRESOLUTION;
   options.scaleTypeValue = 150;
-  
-  let prefix: string;
-  prefix = `${filename.split("\.").shift()}!@#$`;
 
-  let outputFile: File;
-  outputFile = new File(`${destPath}/${filename}`);
-  document.exportForScreens(outputFile, ExportForScreensType.SE_JPEG100, options, undefined, prefix);
+  const prefix: string = `${filename.split('.').shift()}!@#$`;
 
+  const outputFile: File = new File(`${destPath}/${filename}`);
+
+  document.exportForScreens(
+    outputFile, ExportForScreensType.SE_JPEG100, options, undefined, prefix,
+  );
 }
 
 function moveFilesToParentFolder(rootPath: string): void {
   const rootFolder: Folder = new Folder(rootPath);
-  let filesFolder: Folder = new Folder(`${rootPath}/150ppi`);
-  let files: Array<File | Folder>;
-  
-  files = filesFolder.getFiles("");
+  const filesFolder: Folder = new Folder(`${rootPath}/150ppi`);
+  const files: Array<File | Folder> = filesFolder.getFiles('');
 
-  for (let i = files.length; i>0 ;i--) {
-    let file: File | Folder = files.pop();
-    if (file.exists) {
-      file.copy(`${rootFolder}/${file.displayName}`);
-      file.remove();
+  for (let i = files.length; i > 0; i -= 1) {
+    if (files.pop() instanceof File) {
+      const file: File = files.pop();
+      if (file.exists) {
+        file.copy(`${rootFolder}/${file.displayName}`);
+        file.remove();
+      }
     }
   }
   filesFolder.remove();
@@ -70,14 +70,12 @@ function moveFilesToParentFolder(rootPath: string): void {
 
 function correctFileNames(rootPath: string): void {
   const rootFolder: Folder = new Folder(`${rootPath}/150ppi`);
-  let files: Array<File | Folder>;
-  
-  files = rootFolder.getFiles("");
-  
-  for (let i = 0; i<files.length; i++) {
-    let file: File | Folder = files[i];
+  const files: Array<File | Folder> = rootFolder.getFiles('');
+
+  for (let i = 0; i < files.length; i += 1) {
+    const file: File | Folder = files[i];
     if (file.exists) {
-      let newName = `${file.displayName.split("!@#$").shift()}-${i>9 ? "" : "0"}${i}.jpg`;
+      const newName = `${file.displayName.split('!@#$').shift()}-${i > 9 ? '' : '0'}${i}.jpg`;
       file.rename(newName);
     }
   }
@@ -85,12 +83,12 @@ function correctFileNames(rootPath: string): void {
 
 function main(): void {
   const activeDocument: Document = getActiveDocument(app);
-  let destPath: string = getDestinationPath(activeDocument.path.absoluteURI);
+  const destPath: string = getDestinationPath(activeDocument.path.absoluteURI);
 
   createImages(activeDocument, destPath, activeDocument.fullName.displayName);
   correctFileNames(destPath);
   moveFilesToParentFolder(destPath);
-  displayMessage("Files saved");
+  displayMessage('Files saved');
 }
 
 main();
